Develop

+++ VSProjects/API-Sistema-Central/API-Sistema-Central/Services/SubAluguerService.cs
@@ -5,30 +5,65 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using API_Sistema_Central.DTOs;
+using API_Sistema_Central.Models;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 
 namespace API_Sistema_Central.Services
 {
+    public interface ISubAluguerService
+    {
+        public Task<ActionResult<IEnumerable<SubAluguerDTO>>> GetByNifAsync(string nif);
+        public Task<ActionResult<SubAluguerDTO>> GetByIdAsync(int id);
+        public Task<SubAluguerDTO> PostSubAluguerAsync(SubAluguerDTO subAluguerDTO);
+        public Task DeleteSubAluguerAsync(int id);
+    }
+
     public class SubAluguerService : ISubAluguerService
     {
+        private readonly IReservaService _reservaService;
+        private readonly UserManager<Utilizador> _userManager;
+
+        public SubAluguerService(IReservaService reservaService, UserManager<Utilizador> userManager)
+        {
+            _reservaService = reservaService;
+            _userManager = userManager;
+        }
+
         public async Task<ActionResult<IEnumerable<SubAluguerDTO>>> GetByNifAsync(string nif)
         {
+            var u = await _userManager.FindByIdAsync(nif);
+            if (u == null)
+            {
+                throw new Exception("O utilizador não existe.");
+            }
             try
             {
                 IEnumerable<SubAluguerDTO> result;
                 using (HttpClient client = new HttpClient())
                 {
-                    string endpoint = "https://localhost:5005/api/lugares/all/" + nif;
+                    client.DefaultRequestHeaders.Add("Ocp-Apim-Subscription-Key", "aee2fb676a2e4b25a819af617eb64174");
+                    string endpoint = "https://jakim-api-management.azure-api.net/sub-alugueres/api/lugares/all/" + nif;
                     var response = await client.GetAsync(endpoint);
                     response.EnsureSuccessStatusCode();
                     result = await response.Content.ReadAsAsync<IEnumerable<SubAluguerDTO>>();
+                }
+                if (!result.Any())
+                {
+                    return result.ToList();
+                }
+                foreach (SubAluguerDTO s in result)
+                {
+                    bool isReservado = await IsReservado(s.Id);
+                    s.IsReservado = isReservado;
+                    s.NomeParque = await GetParqueNomeByID(s.ParqueId);
                 }
                 return result.ToList();
             }
             catch (Exception)
             {
-                throw new Exception("Não existem lugares associados a este NIF.");
+                throw new Exception("Ligação aos Sub-Alugueres falhou.");
             }
         }
 
@@ -39,11 +74,15 @@ namespace API_Sistema_Central.Services
                 SubAluguerDTO result;
                 using (HttpClient client = new HttpClient())
                 {
-                    string endpoint = "https://localhost:5005/api/lugares/" + id;
+                    client.DefaultRequestHeaders.Add("Ocp-Apim-Subscription-Key", "aee2fb676a2e4b25a819af617eb64174");
+                    string endpoint = "https://jakim-api-management.azure-api.net/sub-alugueres/api/lugares/" + id;
                     var response = await client.GetAsync(endpoint);
                     response.EnsureSuccessStatusCode();
                     result = await response.Content.ReadAsAsync<SubAluguerDTO>();
                 }
+                bool isReservado = await IsReservado(result.Id);
+                result.IsReservado = isReservado;
+                result.NomeParque = await GetParqueNomeByID(result.ParqueId);
                 return result;
             }
             catch (Exception)
@@ -54,13 +93,16 @@ namespace API_Sistema_Central.Services
 
         public async Task<SubAluguerDTO> PostSubAluguerAsync(SubAluguerDTO subAluguerDTO)
         {
+            await ValidarSubAluguer(subAluguerDTO);
+            subAluguerDTO.ParqueId = await GetParqueIdByNome(subAluguerDTO.NomeParque);
             try
             {
                 SubAluguerDTO lugar;
                 using (HttpClient client = new HttpClient())
                 {
+                    client.DefaultRequestHeaders.Add("Ocp-Apim-Subscription-Key", "aee2fb676a2e4b25a819af617eb64174");
                     StringContent content = new StringContent(JsonConvert.SerializeObject(subAluguerDTO), Encoding.UTF8, "application/json");
-                    string endpoint = "https://localhost:5005/api/lugares";
+                    string endpoint = "https://jakim-api-management.azure-api.net/sub-alugueres/api/lugares";
                     var response = await client.PostAsync(endpoint, content);
                     response.EnsureSuccessStatusCode();
                     lugar = await response.Content.ReadAsAsync<SubAluguerDTO>();
@@ -75,11 +117,16 @@ namespace API_Sistema_Central.Services
 
         public async Task DeleteSubAluguerAsync(int id)
         {
+            if (IsReservado(id).Result)
+            {
+                throw new Exception("Proibido: o lugar está reservado.");
+            }
             try
             {
                 using (HttpClient client = new HttpClient())
                 {
-                    string endpoint = "https://localhost:5005/api/lugares/" + id;
+                    client.DefaultRequestHeaders.Add("Ocp-Apim-Subscription-Key", "aee2fb676a2e4b25a819af617eb64174");
+                    string endpoint = "https://jakim-api-management.azure-api.net/sub-alugueres/api/lugares/" + id;
                     var response = await client.DeleteAsync(endpoint);
                     response.EnsureSuccessStatusCode();
                 }
@@ -87,6 +134,95 @@ namespace API_Sistema_Central.Services
             catch
             {
                 throw new Exception("O cancelamento do lugar para Sub-Aluguer falhou.");
+            }
+        }
+
+        private async Task<bool> IsReservado(int lugarId)
+        {
+            bool isReservado = false;
+            var reservas = new List<ReservaAPIParqueDTO>();
+            using (HttpClient client = new HttpClient())
+            {
+                client.DefaultRequestHeaders.Add("Ocp-Apim-Subscription-Key", "aee2fb676a2e4b25a819af617eb64174");
+                string endpoint = "https://jakim-api-management.azure-api.net/sub-alugueres/api/reservas";
+                var response = await client.GetAsync(endpoint);
+                response.EnsureSuccessStatusCode();
+                reservas = await response.Content.ReadAsAsync<List<ReservaAPIParqueDTO>>();
+            }
+            var r = reservas.Where(x => x.LugarId == lugarId);
+            if (r.Any())
+            {
+                isReservado = true;
+            }
+            return isReservado;
+        }
+        private async Task ValidarSubAluguer(SubAluguerDTO subAluguerDTO)
+        {
+            DetalheReservaDTO r = await _reservaService.GetByIdAsync(subAluguerDTO.ReservaSistemaCentralId);
+            if (r == null)
+            {
+                throw new Exception("Não existe reserva para sub-alugar.");
+            }
+            if (r.IsSubAlugado == true)
+            {
+                throw new Exception("Proibido: Este lugar já está Sub-Alugado.");
+            }
+            if (r.NifProprietario != subAluguerDTO.NifProprietario)
+            {
+                throw new Exception("Não é o proprietário dessa reserva.");
+            }
+            if (r.Inicio != subAluguerDTO.Inicio || r.Fim != subAluguerDTO.Fim || r.Andar != subAluguerDTO.Andar || r.Fila != subAluguerDTO.Fila || r.NumeroLugar != subAluguerDTO.Numero)
+            {
+                throw new Exception("Alguns detalhes do sub-aluguer não estão coerentes.");
+            }
+            if (subAluguerDTO.Preco <= 0)
+            {
+                throw new Exception("O preço não pode ser igual ou menor que zero.");
+            }
+        }
+        private static async Task<int> GetParqueIdByNome(string nome)
+        {
+            try
+            {
+                ParqueDTO p;
+                using (HttpClient client = new HttpClient())
+                {
+                    client.DefaultRequestHeaders.Add("Ocp-Apim-Subscription-Key", "aee2fb676a2e4b25a819af617eb64174");
+                    string endpoint1 = "https://jakim-api-management.azure-api.net/sub-alugueres/api/parques/";
+                    var response1 = await client.GetAsync(endpoint1);
+                    response1.EnsureSuccessStatusCode();
+                    var lista = await response1.Content.ReadAsAsync<List<ParqueDTO>>();
+                    p = lista.First(t => t.Rua == nome);
+                }
+                if (p == null)
+                {
+                    throw new Exception("Este parque não existe.");
+                }
+                return p.Id;
+            }
+            catch (Exception)
+            {
+                throw new Exception("GetParqueIdByNome() falhou.");
+            }
+        }
+        private static async Task<string> GetParqueNomeByID(int id)
+        {
+            try
+            {
+                ParqueDTO p;
+                using (HttpClient client = new HttpClient())
+                {
+                    client.DefaultRequestHeaders.Add("Ocp-Apim-Subscription-Key", "aee2fb676a2e4b25a819af617eb64174");
+                    string endpoint1 = "https://jakim-api-management.azure-api.net/sub-alugueres/api/parques/" + id;
+                    var response1 = await client.GetAsync(endpoint1);
+                    response1.EnsureSuccessStatusCode();
+                    p = await response1.Content.ReadAsAsync<ParqueDTO>();
+                }
+                return p.Rua;
+            }
+            catch (Exception)
+            {
+                throw new Exception("GetParqueNomeByID() falhou.");
             }
         }
     }
